requirements.txt
Flask==3.0.0
APScheduler==3.10.4
PyYAML==6.0.1
pytz==2023.3
Werkzeug==3.0.1

gunicorn==21.2.0
